fix: skip da client when mock backend
fix: download genesis file right after rollapp init
refactor: move genesis funcs to utils
